first_release_of_security_system

+++ app/Actions/UserActions.php
@@ -2,6 +2,7 @@
 
 namespace MasterStudents\Actions;
 
+use MasterStudents\Core\Hash;
 use MasterStudents\Models\User;
 
 trait UserActions
@@ -17,6 +18,8 @@ trait UserActions
     ];
 
     try {
+      $data["password"] = Hash::make($data["password"]);
+
       $user = User::create($data);
     } catch (\Exception $e) {
       return (object) [
@@ -41,8 +44,39 @@ trait UserActions
       "validator" => $validator
     ];
 
+
     try {
-      $user = User::find($id)->update($data);
+      $user = User::find($id)->update(map($data)->filterWithKey(fn ($v, $k) => $k != "password")->toArray());
+    } catch (\Exception $e) {
+      return (object) [
+        "server_error" => true
+      ];
+    }
+
+    return (object) [
+      "server_error" => false,
+      "updated" => true,
+      "user" => $user
+    ];
+  }
+
+  public static function changePasswordAction(array $data, $id)
+  {
+    if (is_null(User::find($id))) return (object) [
+      "server_error" => true
+    ];
+
+    $validator = request()->validate(User::changePasswordRules(User::find($id)));
+
+
+    if ($validator->fails()) return (object) [
+      "server_error" => false,
+      "updated" => false,
+      "validator" => $validator
+    ];
+
+    try {
+      $user = User::find($id)->update(["password" => Hash::make($data["password"])]);
     } catch (\Exception $e) {
       return (object) [
         "server_error" => true
@@ -59,7 +93,9 @@ trait UserActions
   public static function deleteAction($id)
   {
     try {
-      User::find($id)->delete();
+      if (!is_null(User::find($id))) {
+        User::find($id)->delete();
+      }
     } catch (\Exception $e) {
       return (object) [
         "server_error" => true
